backend/dropa_app/views.py
from rest_framework.views import APIView
from rest_framework.response import Response
from rest_framework import status
from .models import *
from .serializers import *
import pyotp

class PredictDeliveryTimeView(APIView):
    def post(self, request):
        # Integrate ML model here
        return Response({'eta': 'Predicted ETA'}, status=status.HTTP_200_OK)

class AnomalyDetectionView(APIView):
    def post(self, request):
        # Integrate anomaly detection model here
        return Response({'anomaly': 'Detected anomaly'}, status=status.HTTP_200_OK)

class ForecastView(APIView):
    def get(self, request):
        # Integrate forecasting model here
        return Response({'forecast': 'Forecasted volume'}, status=status.HTTP_200_OK)

class SendOTPView(APIView):
    def post(self, request):
        package_id = request.data.get('package_id')
        otp = pyotp.TOTP(pyotp.random_base32()).now()
        OTPLog.objects.create(package_id=package_id, otp_code=otp)
        return Response({'otp': otp}, status=status.HTTP_200_OK)

class VerifyOTPView(APIView):
    def post(self, request):
        package_id = request.data.get('package_id')
        otp_code = request.data.get('otp_code')
        otp_log = OTPLog.objects.filter(package_id=package_id, otp_code=otp_code).first()
        if otp_log:
            otp_log.verified = True
            otp_log.save()
            return Response({'verified': True}, status=status.HTTP_200_OK)
        return Response({'verified': False}, status=status.HTTP_400_BAD_REQUEST)

class DropaBotView(APIView):
    def post(self, request):
        # Integrate RAG chatbot here
        return Response({'answer': 'Chatbot response'}, status=status.HTTP_200_OK)

class PackageListView(APIView):
    def get(self, request):
        packages = Package.objects.all()
        serializer = PackageSerializer(packages, many=True)
        return Response(serializer.data)

class CourierStatsView(APIView):
    def get(self, request):
        # Return courier stats
        return Response({'stats': 'Courier stats'}, status=status.HTTP_200_OK)

class DashboardView(APIView):
    def get(self, request):
        # Return dashboard data
        return Response({'dashboard': 'Dashboard data'}, status=status.HTTP_200_OK)

class LoginView(APIView):
    def post(self, request):
        # Implement login logic
        return Response({'login': 'User logged in'}, status=status.HTTP_200_OK)

class RegisterView(APIView):
    def post(self, request):
        # Implement registration logic
        return Response({'register': 'User registered'}, status=status.HTTP_201_CREATED)
